cuda_toolkit: EUPS support for CUDA toolkit
CUDA Toolkit is a third-party GPU C/C++ compiler that we can't redistribute.
But we can make an EUPS wrapper for it, that will set the appropriate
environment variables and symlinks so packages with GPU acceleration know
where to find it.

With this package, to enable GPU builds the user must do:

	export CUDA=/usr/local/cuda
	eups distrib install cuda_toolkit 4.1

The .bld script checks for $CUDA, then for the compiler in $CUDA/bin, and
then makes sure the installed version and the package version ($release)
are the same. It then creates the apropriate symlinks to the compiler.